tensorrt/python_poc/requirements.txt
matplotlib==3.3.3
numpy==1.19.4
torch==1.7.0
opencv_python==4.4.0.46
pandas==1.1.4
scipy==1.5.4
tqdm==4.53.0
torchvision==0.8.1
coremltools==4.0
onnx==1.8.0
Pillow==8.0.1
PyYAML==5.3.1
seaborn==0.11.0
thop==0.0.31-2005241907
